Extract CylinderItem to separate file
